Unwrap StatsCollector at get_stats
Converting a StatsCollector to a StatsReport takes ownership of the collector.
With the From trait implemented on <Arc<Mutex<StatsCollector>>, it's possible
that the caller may be changed such that a reference to the Arc may still be
open somewhere else.

This moves the unwrapping of the Arc/Mutex to the place where stats are collected,
so that any future code changing a StatsCollector to a StatsReport is in charge
of transfering ownership correctly.