One single direction per extmap
SDP has the syntax: a=extmap:<id>/<direction> <uri>

The /<direction> part is optional (and rarely seen it in the wild).

The current MediaEngine API makes it seem like it's possible to
provide many directions for each map (it takes a Vec). This commit
changes that sov only single optional direction can be provided.

This simplifies internal matching logic and fixes a bug where setting
the transceiver direction `inactive`, would (incorrectly) result in no
extmaps matching for the m-line.