sctp: reduce AI lock contention (#363)
As discussed in #360 the lock on the internal association is very contended in high send-bandwidth situations. This PR achieves two things:

1. Pull the marshalling of packets outside of the critical section thus reducing the time the lock is taken by the write loop
2. Schedule the marshalling and sending as a new tokio::task::Task which makes tokio schedule this on another thread allowing the read loop to make progress in parallel while the write loop is working on that
This in itself does not really increase the bandwidth but with improvements to the marshalling code itself gains can be head now, that were previously (at least partially) blocked by the lock contention.

It should also improve situations where both sides send a lot of data because then the write and read loops would both be very busy and fighting for the lock even more.