fix understated `ipnet` version requirement (#521)
`webrtc-util` uses `IpNet::with_netmask`, which was introduced in
version 2.6.0. If someone is using a `Cargo.lock` with an older version of
`ipnet`, they'll get a compile error when introducing `webrtc-util` (or
upgrading it to after this call was added). Adding the correct version
requirement in `Cargo.toml` forces the upgrade to fix this.