interceptor: use futures::future::join_all in stats interceptor (#710)
The previous implementation iterated through a list of futures and awaited them one by one. This is inefficient as it doesn't allow the futures to run concurrently.

This commit refactors the code to use `futures::future::join_all` to await all futures concurrently, which is more efficient. This also involved adding the `futures` crate as a dependency to the `interceptor` crate.

Signed-off-by: Xiaobo Liu <cppcoffee@gmail.com>