[ice] Avoid unchecked incrementation of port number to avoid potential panic due to overflow in debug, or silent overflow in release (#719)
When `port_start` is assigned the value `u16::MAX`, and the socket
cannot be bound to port `u16::MAX`, an unhandled overflow will occur.

When building webrtc-rs using the debug profile, this overflow will lead
to a panic, and worse yet in the release profile, a silent overflow and
wrap around will occur, allowing `listen_udp_in_port_range` to bind to a
port outside of the specified port range.