[datachannel] Add the `max_message_size` config field into the datachannel's config
Given that we are no longer fixed to the 64KB max size for message
sizes, the user needs a way to know the actual max message size a
datachannel may send before returning an error. This information is
particularly useful for applications where large messages are chunked
and framed in the application itself.